feat(ui): ARC Design System — reator nativo, fontes e build funcional
Substitui a UI da FASE 7 pelo ARC: identidade própria de HUD radial,
implementada em SkiaSharp e rodando nativamente no MAUI.

Design System
- docs/design-system/: 8 documentos cobrindo tokens, tipografia, 24
  componentes, navegação radial, movimento, guia MAUI e checklists.
- Tokens em C# (ui-engine/Design) e XAML (Resources/Styles), com teste
  de paridade descrito em 07-CHECKLISTS §3.3.
- InkLo recalibrado de 0.42 para 0.66 de alpha: media 2.48:1 e reprovava
  em WCAG AA. Agora 4.73:1.

Reator
- ReactorRenderer (SkiaSharp): bisel de metal escovado, 3 anéis de placas
  biseladas com luz escapando pelas frestas, 8 bobinas com pulso viajante,
  garganta de lentes e núcleo branco-quente com bloom e flare.
- Varredura especular independente do estado: o metal reluz mesmo offline,
  cumprindo "nada fica parado" sem fazer sistema morto parecer vivo.
- Estados dirigidos por ArcBrain a partir do ciclo de vida do assistente.
  O chip de status reflete, nunca comanda.

Tipografia
- Rajdhani e JetBrains Mono (SIL OFL 1.1) empacotadas com suas licenças.
- 155 ocorrências de Courier New substituídas pelo alias ArcData.

Correções de build (o app nunca compilou nem abriu)
- Grpc.Tools engancha em BeforeCompile, inalcançável no Windows: o WinUI
  compila antes via XamlPreCompile e abortava. Codegen movido para
  BeforeTargets="XamlPreCompile;CoreCompile".
- GrpcClient reescrito: namespaces SextaFeira.*.V1, enum RelationType com
  prefixo, e seção de voz refeita sobre o StreamVoice bidirecional real —
  os quatro RPCs unários que ela chamava não existem no proto.
- DefaultEngine deixou de ser sealed; 15 Pages passaram a declarar
  HudBasePage como raiz do XAML; PluginManifest e MemorySearchResult
  duplicados resolvidos; serviços passaram a satisfazer I*Engine.
- MainPage resolvia o ViewModel via Handler no construtor, onde ele é
  sempre null — trocado por injeção. Era o crash de startup (0xC000027B).
- WindowsAppSDKSelfContained: a máquina precisa do runtime 1.5 exato.
- Logger de crash para exceções stowed do WinUI, que não passam pelo
  AppDomain e não deixam stack no Event Log.

Verificado: Windows e Android compilam com 0 erros e 0 avisos; o app abre
e renderiza o ARC, com transições de estado automáticas.

+++ apps/maui/CognitiveHUD/Models/CognitionModels.cs
@@ -15,7 +15,9 @@ public record HealthStatus(
     string Version,
     bool OllamaOnline,
     bool VoiceAvailable,
-    int UptimeSeconds);
+    // int64 on the wire — the proto reports uptime as a 64-bit count and
+    // narrowing it here is what broke both call sites.
+    long UptimeSeconds);
 
 /// <summary>Conversation state for tracking multi-turn chat.</summary>
 public record Conversation(
